+++ sandbox-backend/src/main/java/com/exadel/team2/sandbox/configuration/SecurityConfig2.java
@@ -1,0 +1,50 @@
+package com.exadel.team2.sandbox.configuration;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.NoOpPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+import javax.sql.DataSource;
+
+@Configuration
+@EnableWebSecurity
+public class SecurityConfig2 extends WebSecurityConfigurerAdapter {
+
+    final DataSource dataSource;
+
+    public SecurityConfig2(DataSource dataSource) {
+        this.dataSource = dataSource;
+    }
+//todo                 тут пробую зробити з зашифрований  варыант з бд
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+
+
+    
+
+    @Override
+    protected void configure(AuthenticationManagerBuilder auth) throws Exception {
+        auth.jdbcAuthentication().dataSource(dataSource);
+//                .passwordEncoder(passwordEncoder());
+    }
+
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http.authorizeRequests()
+                .antMatchers("api/event/5").hasAnyRole("DEVELOPER", "ADMIN")
+                .antMatchers("/event/all").permitAll()
+                .antMatchers("/event-type").hasRole("Super-admin")
+                .and().formLogin();
+    }          //доступ к страницам - каким ролям! Працюэ але незашифровано! поки
+
+
+}
