+++ sandbox-web/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-server.port=8083
+server.port=8090
 spring.application.name=sandbox
 server.servlet.context-path=/api
 server.error.include-message=always
